+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 * [BUGFIX] Querier: fixed panic when querying exemplars and using `-distributor.shard-by-all-labels=false`. #4473
 * [BUGFIX] Querier: honor querier minT,maxT if `nil` SelectHints are passed to Select(). #4413
 * [BUGFIX] Compactor: fixed panic while collecting Prometheus metrics. #4483
+* [BUGFIX] Ring: extend replication when instance heartbeat timeoutgst. #4493
 
 
 ## 1.10.0 / 2021-08-03